+++ lib/fluent_ui.dart
@@ -24,7 +24,10 @@ export 'package:flutter/material.dart'
         TooltipVisibility,
         TooltipTriggerMode,
         TextInputAction,
-        MaterialLocalizations;
+        MaterialLocalizations,
+        TextSelectionTheme,
+        TextSelectionThemeData,
+        SelectableText;
 export 'package:scroll_pos/scroll_pos.dart';
 
 export 'src/app.dart';
@@ -84,6 +87,7 @@ export 'src/controls/form/pickers/date_picker.dart';
 export 'src/controls/form/pickers/time_picker.dart';
 export 'src/controls/form/text_form_box.dart';
 export 'src/controls/form/form_row.dart';
+export 'src/controls/form/selection_controls.dart';
 
 export 'src/styles/motion/page_transitions.dart';
 export 'src/styles/acrylic.dart';